requirements.txt
networkx>=3.0
pydot>=1.4.2
matplotlib>=3.5
Pillow>=9.0
numpy>=1.23
pandas>=1.5
openpyxl>=3.0
